Created a separate privilege checker function for create new user and database 3 backup
Since new users don't have data saved in the database, getting the role level using id_no will fail. Instead, a new function that checks the role_level using the role_id (the role_id that the user will be set to) is created and connected to the create user function.