Wire the Open Button in Syllabus Templates to RTEditor
After selecting a template tile and either double-clicking, clicking and press enter, or clicking a tile and clicking the open button now opens the RTEditor in a new tab.
Whether it renders the correct template content or not is unknown at the moment because the templates are empty. It is recommended to finish the save/update feature to test this.