Update .gitignore
Updated .gitignore to ignore packages and other environment variables. These packages and environment variables should be setup on the individual devices for the app to work i.e. every device has it's own unique implementation of these files so they don't need to be pushed instead, they need to be installed and set up on the individual devices.