Updated RTEditor Controller and View
Updated the RTEditor controller and views to align with the new Node.js and Vite setup. Deprecated the importmap because Vite is now the one responsible for importing necessary packages. Also updated the view /app/Modules/RTEditor/Views/index.php to use the APP_ENV environment variable to decide whether to use the dev path or prod path. Also changed the markup for the new editor flow.